adjusting parametres in primer3, repeating step1 if primer design unsuccessful, fixing errors, config file, input file

+++ fastGENdesigner3-primerpooler.sh
@@ -1,21 +1,60 @@
 #!/bin/bash
-input_file=$1
-output_folder=$2
+
+while [[ $# -gt 0 ]]; do
+ case $1 in
+    --input)
+       input_file="$2"
+       shift
+       shift
+       ;;
+    --output)
+       output_folder="$2"
+       shift
+       shift
+       ;;
+    --pools)
+       pools_num="$2"
+       shift
+       shift
+       ;;
+    *)
+     echo "Input not understood."
+       exit 1
+       ;;
+ esac
+done
+
+if [ "${#input_file}" == "0" ]; then input_file=$(cat fastGENdesigner-input | grep -v '^#' | grep "input_file" | cut -d '=' -f 2); fi
+if [ "${#output_folder}" == "0" ]; then output_folder=$(cat fastGENdesigner-input | grep -v '^#' | grep "output_dir" | cut -d '=' -f 2); fi
+if [ "${#pools_num}" == "0" ]; then pools_num=$(cat fastGENdesigner-input | grep -v '^#' | grep "pools" | cut -d '=' -f 2); fi
+primer_pooler=$(cat config | grep "primer_pooler" | cut -d '=' -f 2)
+
+temp_file=$(mktemp)
+exec > >(tee "$temp_file") 2>&1
 
 echo "Starting primer pooler"
-
 # NUMBER OF POOLS - SUGGESTION
-~/pooler/pooler --suggest-pools --genome=/home/ppola/pooler/hg38.2bit $output_folder/primers.fasta 2> ${output_folder}/pooler_output.txt
-pools_num=$(tail -1 ${output_folder}/pooler_output.txt | tr -dc '0-9' )
-pools_num=$((pools_num-1)) # zmensit o jedna - prediskutovat
-rm ${output_folder}/pooler_output.txt
-echo "Number of pools: $pools_num"
+if [ "${#pools_num}" == "0" ]
+then
+	$primer_pooler --suggest-pools --genome=fastGENdesigner_files/hg38.2bit $output_folder/primers.fasta 2> ${output_folder}/pooler_output.txt
+	pools_num=$(tail -1 ${output_folder}/pooler_output.txt | tr -dc '0-9' )
+	#pools_num=$((pools_num-1)) # zmensit o jedna - prediskutovat
+	rm ${output_folder}/pooler_output.txt
+	rm overlap-report*.txt
+	echo "Number of suggested pools: $pools_num"
+else
+	echo "Number of given pools: $pools_num"
+fi
+
+rm -rf ${output_folder}/pooler_output
+mkdir -p ${output_folder}/pooler_output
 
 # PRIMER POOLER
-~/pooler/pooler --pools=$pools_num,1,$output_folder/poolfile --genome=/home/ppola/pooler/hg38.2bit --seedless $output_folder/primers.fasta 2> ${output_folder}/pooler_output.txt
-mv overlap-report-1.txt $output_folder
+#$primer_pooler --pools=$pools_num,1,$output_folder/pooler_output/poolfile --genome=fastGENdesigner_files/hg38.2bit --seedless $output_folder/primers.fasta 2> ${output_folder}/pooler_output/pooler_output.txt
+$primer_pooler --pools=$pools_num,1,$output_folder/pooler_output/poolfile --genome=fastGENdesigner_files/hg38.2bit $output_folder/primers.fasta 2> ${output_folder}/pooler_output/pooler_output.txt
+rm overlap-report*.txt 
 
-echo "You can find poolfiles and primerpooler output in: $output_folder"
+echo "Poolfiles and pooler output created"
 echo ""
 
 # POOLS DISTRIBUTION
@@ -23,14 +62,16 @@ good_pools=$(Rscript pooler_summary.R "input_file='$input_file'" "output_folder=
 
 if [ "${#good_pools}" != "0" ]
 then
-	echo "I have chosen ${good_pools}. Starting to score them."
-	echo "Primer pooler scoring" > ${output_folder}/poolfiles_score.txt
-
+	echo "Selected pools:  ${good_pools}. Initiating scoring process"
+	echo "Primer pooler scoring" > ${output_folder}/pooler_output/poolfiles_score.txt
+	
+	temp_file1=$(mktemp)
 	for pool in $good_pools; do
-		echo $pool >> ${output_folder}/poolfiles_score.txt
-		~/pooler/pooler ${output_folder}/${pool} --print-bonds=5 >> ${output_folder}/poolfiles_score.txt 2> ${output_folder}/pooler_output.txt; 
+		echo $pool >> ${output_folder}/pooler_output/poolfiles_score.txt
+		$primer_pooler ${output_folder}/pooler_output/${pool} --print-bonds=5 >> ${output_folder}/pooler_output/poolfiles_score.txt 2> ${temp_file1}; 
 	done
-
+	rm "$temp_file1"
+	
 	scores=$(Rscript pooler_scoring.R "input_file='$input_file'" "output_folder='$output_folder'")
 	read -ra scores_array <<< "$scores"
 	read -ra pools_array <<< "$good_pools"
@@ -45,4 +86,5 @@ then
 fi	
 echo "Done"
 
-
+cat "$temp_file" >> ${output_folder}/output.log
+rm "$temp_file"
